Run everything inside container
